+++ macros/src/removable_derive.rs
@@ -31,7 +31,7 @@ pub fn impl_removable_derive(ast: &syn::DeriveInput) -> Result<proc_macro2::Toke
                 use actix::prelude::*;
 
                 let widget = &#widget_to_remove;
-                if let Some(parent) = widget.get_parent() {
+                if let Some(parent) = widget.parent() {
                     let parent = parent.downcast::<gtk::Container>().unwrap();
                     let widget = widget.clone();
                     ctx.stop();