+++ src/prop_sync.rs
@@ -44,7 +44,7 @@ impl GetProps for gtk::Entry {
     type GetterType = String;
 
     fn get_props(&self) -> Self::GetterType {
-        self.get_text().to_string()
+        self.text().to_string()
     }
 }
 
@@ -60,6 +60,6 @@ impl GetProps for gtk::CheckButton {
     type GetterType = bool;
 
     fn get_props(&self) -> Self::GetterType {
-        self.get_active()
+        self.is_active()
     }
 }