+++ tests/recusive_widgets_from_builder.rs
@@ -41,25 +41,25 @@ fn test_recusive_widgets_from_builder() -> anyhow::Result<()> {
     let flat_widgets: FlatWidgets = bld.widgets()?;
     let grouped_widgets: GroupedWidgets = bld.widgets()?;
 
-    assert!(grouped_widgets.group_a.text1.get_text().is_empty());
-    assert!(grouped_widgets.group_a.text2.get_text().is_empty());
-    assert!(grouped_widgets.group_b.text3.get_text().is_empty());
-    assert!(grouped_widgets.group_b.text4.get_text().is_empty());
+    assert!(grouped_widgets.group_a.text1.text().is_empty());
+    assert!(grouped_widgets.group_a.text2.text().is_empty());
+    assert!(grouped_widgets.group_b.text3.text().is_empty());
+    assert!(grouped_widgets.group_b.text4.text().is_empty());
 
     flat_widgets.text1.set_text("Text 1");
     flat_widgets.text2.set_text("Text 2");
     flat_widgets.text3.set_text("Text 3");
     flat_widgets.text4.set_text("Text 4");
 
-    wait_for!(flat_widgets.text1.get_text() == "Text 1")?;
-    wait_for!(flat_widgets.text2.get_text() == "Text 2")?;
-    wait_for!(flat_widgets.text3.get_text() == "Text 3")?;
-    wait_for!(flat_widgets.text4.get_text() == "Text 4")?;
+    wait_for!(flat_widgets.text1.text() == "Text 1")?;
+    wait_for!(flat_widgets.text2.text() == "Text 2")?;
+    wait_for!(flat_widgets.text3.text() == "Text 3")?;
+    wait_for!(flat_widgets.text4.text() == "Text 4")?;
 
-    assert!(grouped_widgets.group_a.text1.get_text() == "Text 1");
-    assert!(grouped_widgets.group_a.text2.get_text() == "Text 2");
-    assert!(grouped_widgets.group_b.text3.get_text() == "Text 3");
-    assert!(grouped_widgets.group_b.text4.get_text() == "Text 4");
+    assert!(grouped_widgets.group_a.text1.text() == "Text 1");
+    assert!(grouped_widgets.group_a.text2.text() == "Text 2");
+    assert!(grouped_widgets.group_b.text3.text() == "Text 3");
+    assert!(grouped_widgets.group_b.text4.text() == "Text 4");
 
     Ok(())
 }