+++ src/event_loops_bridge.rs
@@ -36,7 +36,7 @@ pub fn try_block_on<F: Future>(fut: F) -> Result<<F as Future>::Output, F> {
 pub fn run_actix_inside_gtk_event_loop() -> std::io::Result<glib::SourceId> {
     let source_id = glib::idle_add(|| {
         try_block_on(async {
-            actix::clock::sleep(core::time::Duration::new(0, 0)).await;
+            actix::clock::sleep(core::time::Duration::new(0, 10_000_000)).await;
         })
         .map_err(|_| "`idle_add` function called inside Actix context")
         .unwrap();